move main optimise 2pF model function to optimisation.cpp and resolve includes
